refactor: FSD 아키텍처 원칙에 따라 cart와 coupon 비즈니스 로직을 features에서 entities로 이동
