+++ src/entities/cart/api/updateCartItemQuantity.ts
@@ -1,4 +1,4 @@
-import { UpdateCartItemQuantity } from '@entities/cart/type/cartItem.type';
+import { UpdateCartItemQuantity } from '@entities/cart';
 import { httpClient } from '@shared/api/httpClient';
 
 const ERROR_MESSAGE = '장바구니 상품 수량을 업데이트하는 데 실패했습니다.';
