+++ src/entities/cart/index.ts
@@ -1,6 +1,13 @@
-export type { CartItemType, OrderItemType } from './type/cartItem.type';
+export type { CartItemType, UpdateCartItemQuantity } from './type/cartItem.type';
 export { getCartItems } from './api/getCartItems';
 export { deleteCartItem } from './api/deleteCartItem';
 export { updateCartItemQuantity } from './api/updateCartItemQuantity';
 export { useCartData } from './model/createCartData';
 export { cartActions } from './model/cartActions';
+export { useCart } from './model/useCart';
+export { CartProvider, useCartContext } from './model/provider/CartProvider';
+export {
+  calculateOrderPrice,
+  calculateDeliveryFee,
+  calculatePaymentInfo,
+} from './utils/cartPriceCalculator';
