+++ src/entities/cart/utils/cartPriceCalculator.ts
