+++ src/entities/coupon/constants/couponRule.ts
