+++ src/entities/coupon/utils/formatter.ts
