+++ src/entities/order/type/order.type.ts
@@ -1,0 +1,1 @@
+export type OrderItemIdListType = number[];
