+++ src/widget/cart/CartContent/CartContent.test.tsx
@@ -1,7 +1,7 @@
 import { act, fireEvent, render, screen, waitFor, within } from '@testing-library/react';
 import { MemoryRouter } from 'react-router-dom';
 import CartContent from './CartContent';
-import { CartProvider } from '@/features/cart/model/provider/CartProvider';
+import { CartProvider } from '@entities/cart';
 
 let { mockCartItems } = vi.hoisted(() => {
   return {
